+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Dates are in YYYY-MM-DD format.
 - Fixes `FileManager`'s logic for finding source nodes. This prevents headers from being scanned multiple times.
 - Adds checks so that invalid target names are caught early.
 - Adds default profiles to `sbuildr test`.
+- Fixes a bug where multiple requests to a Dependency for the same library would result in only the last request being valid.
 
 ## v0.6.0 (2019-08-27)
 - Adds `depends` argument to `Project.library()`, `Project.executable()`, `Project.test()`, and `Project.interfaces()`. This can be used to specify any dependencies not captured in the `libs` argument - for example, header-only packages.